third_party/blink/web_tests/platform/mac/external/wpt/html/rendering/widgets/the-select-element/select-as-listbox-default-styles.tentative-expected.txt

This is a testharness.js-based test.
Found 38 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] <select multiple=""><option>1 - border-top-width
  assert_equals: expected "0px" but got "1px"
[FAIL] <select multiple=""><option>1 - border-right-width
  assert_equals: expected "0px" but got "1px"
[FAIL] <select multiple=""><option>1 - border-bottom-width
  assert_equals: expected "0px" but got "1px"
[FAIL] <select multiple=""><option>1 - border-left-width
  assert_equals: expected "0px" but got "1px"
[FAIL] <select multiple=""><option>1 - border-top-style
  assert_equals: expected "none" but got "solid"
[FAIL] <select multiple=""><option>1 - border-right-style
  assert_equals: expected "none" but got "solid"
[FAIL] <select multiple=""><option>1 - border-bottom-style
  assert_equals: expected "none" but got "solid"
[FAIL] <select multiple=""><option>1 - border-left-style
  assert_equals: expected "none" but got "solid"
[FAIL] <select multiple=""><option>1 - border-top-color
  assert_equals: expected "rgb(0, 0, 0)" but got "rgb(118, 118, 118)"
[FAIL] <select multiple=""><option>1 - border-right-color
  assert_equals: expected "rgb(0, 0, 0)" but got "rgb(118, 118, 118)"
[FAIL] <select multiple=""><option>1 - border-bottom-color
  assert_equals: expected "rgb(0, 0, 0)" but got "rgb(118, 118, 118)"
[FAIL] <select multiple=""><option>1 - border-left-color
  assert_equals: expected "rgb(0, 0, 0)" but got "rgb(118, 118, 118)"
[FAIL] <select multiple=""><option>1 - align-items
  assert_equals: expected "normal" but got "flex-start"
[FAIL] <select multiple=""><option>1 - white-space
  assert_equals: expected "normal" but got "nowrap"
[FAIL] <select multiple=""><option>1 - background-color
  assert_equals: expected "rgba(0, 0, 0, 0)" but got "rgb(255, 255, 255)"
[FAIL] <select multiple=""><option>1 - cursor
  assert_equals: expected "auto" but got "default"
[FAIL] <select multiple=""><option>1 - font-size
  assert_equals: expected "16px" but got "13.3333px"
[FAIL] <select multiple=""><option>1 - font-family
  assert_equals: expected "Times" but got "Arial"
[FAIL] <select multiple=""><option>1 - vertical-align
  assert_equals: expected "baseline" but got "text-bottom"
[FAIL] <option>1 (in <select multiple="">) - padding-right
  assert_equals: expected "0px" but got "2px"
[FAIL] <option>1 (in <select multiple="">) - padding-bottom
  assert_equals: expected "0px" but got "1px"
[FAIL] <option>1 (in <select multiple="">) - padding-left
  assert_equals: expected "0px" but got "2px"
[FAIL] <option>1 (in <select multiple="">) - white-space
  assert_equals: expected "normal" but got "nowrap"
[FAIL] <option>1 (in <select multiple="">) - cursor
  assert_equals: expected "auto" but got "default"
[FAIL] <option>1 (in <select multiple="">) - font-size
  assert_equals: expected "16px" but got "13.3333px"
[FAIL] <option>1 (in <select multiple="">) - font-family
  assert_equals: expected "Times" but got "Arial"
[FAIL] <optgroup label="2"><option>3 (in <select multiple="">) - white-space
  assert_equals: expected "normal" but got "nowrap"
[FAIL] <optgroup label="2"><option>3 (in <select multiple="">) - cursor
  assert_equals: expected "auto" but got "default"
[FAIL] <optgroup label="2"><option>3 (in <select multiple="">) - font-weight
  assert_equals: expected "400" but got "700"
[FAIL] <optgroup label="2"><option>3 (in <select multiple="">) - font-size
  assert_equals: expected "16px" but got "13.3333px"
[FAIL] <optgroup label="2"><option>3 (in <select multiple="">) - font-family
  assert_equals: expected "Times" but got "Arial"
[FAIL] <option>3 (in <select multiple=""><optgroup label="2">) - padding-right
  assert_equals: expected "0px" but got "2px"
[FAIL] <option>3 (in <select multiple=""><optgroup label="2">) - padding-bottom
  assert_equals: expected "0px" but got "1px"
[FAIL] <option>3 (in <select multiple=""><optgroup label="2">) - padding-left
  assert_equals: expected "0px" but got "20px"
[FAIL] <option>3 (in <select multiple=""><optgroup label="2">) - white-space
  assert_equals: expected "normal" but got "nowrap"
[FAIL] <option>3 (in <select multiple=""><optgroup label="2">) - cursor
  assert_equals: expected "auto" but got "default"
[FAIL] <option>3 (in <select multiple=""><optgroup label="2">) - font-size
  assert_equals: expected "16px" but got "13.3333px"
[FAIL] <option>3 (in <select multiple=""><optgroup label="2">) - font-family
  assert_equals: expected "Times" but got "Arial"
Harness: the test ran to completion.
